[TEST] Extracts reusable `FeedStoreSpecs` helper methods into a shared scope
Can be used by other `FeedStore` implementation test.
